move msw handlers and server to fake-api package
