+++ packages/kuma-gui/src/app/fake-fs/index.ts
@@ -1,18 +1,31 @@
+import { mswHandlers } from '@kumahq/fake-api/msw'
+
 import type { ServiceDefinition, Token } from '@/services/utils'
 import { token } from '@/services/utils'
-import type { FS } from '@/test-support'
-import { mswHandlers } from '@/test-support'
+import type { EndpointDependencies, FS } from '@/test-support'
+import { dependencies } from '@/test-support/fake'
 
 const $ = {
   fakeFS: token<FS>('fake.fs'),
+  dependencies: token<EndpointDependencies>('dependencies'),
 }
 
 export const services = (app: Record<string, Token>): ServiceDefinition[] => [
+  [$.dependencies, {
+    service: (env) => ({
+      ...dependencies,
+      env,
+    }),
+    arguments: [
+      app.env,
+    ],
+  }],
   [token('fake.msw.handlers'), {
     service: mswHandlers,
     arguments: [
       app.env,
       $.fakeFS,
+      $.dependencies,
     ],
     labels: [
       app.mswHandlers,
