+++ packages/kuma-gui/src/app/hostname-generators/data/index.spec.ts
@@ -2,6 +2,7 @@ import { describe, expect, test as _test } from 'vitest'
 
 import { HostnameGenerator } from './HostnameGenerator'
 import { plugin, server } from '@/test-support/data'
+import { dependencies } from '@/test-support/fake'
 import mock from '@/test-support/mocks/src/hostname-generators/_/_overview'
 
 describe('HostnameGenerator', () => {
@@ -11,7 +12,7 @@ describe('HostnameGenerator', () => {
       params: {
         name: 'local-mesh-external-service',
       },
-    }),
+    }, dependencies),
   ))
 
   describe('hostnameGenerator.$raw', () => {
