+++ packages/kuma-gui/src/app/zone-ingresses/data/index.spec.ts
@@ -2,6 +2,7 @@ import { describe, expect, test as _test } from 'vitest'
 
 import { ZoneIngressOverview } from './'
 import { plugin, server } from '@/test-support/data'
+import { dependencies } from '@/test-support/fake'
 import mock from '@/test-support/mocks/src/zone-ingresses/_/_overview'
 
 // zoneIngress tests are very similar to zoneEgress tests
@@ -19,7 +20,7 @@ describe('ZoneIngressOverview', () => {
       params: {
         name: 'zone',
       },
-    }),
+    }, dependencies),
   ))
   //
   describe('labels', () => {
