+++ packages/kuma-gui/src/app/zones/data/index.spec.ts
@@ -2,6 +2,7 @@ import { describe, expect, test as _test } from 'vitest'
 
 import { ZoneOverview } from './'
 import { plugin, server } from '@/test-support/data'
+import { dependencies } from '@/test-support/fake'
 import mock from '@/test-support/mocks/src/zones/_/_overview'
 
 describe('ZoneOverview', () => {
@@ -11,7 +12,7 @@ describe('ZoneOverview', () => {
       params: {
         name: 'zone',
       },
-    }),
+    }, dependencies),
   ))
   //
   describe('zoneInsight.config', () => {
